backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pip>=23.0",
    "camel-ai==0.2.90a5",
    "fastapi>=0.115.12",
    "fastapi-babel>=1.0.0",
    "uvicorn[standard]>=0.34.2",
    "pydantic-i18n>=0.4.5",
    "python-dotenv>=1.1.0",
    "httpx[socks]>=0.28.1",
    "pydash>=8.0.5",
    "inflection>=0.5.1",
    "aiofiles>=24.1.0",
    "openai>=1.99.3,<2",
    "nodejs-wheel>=22.18.0",
    "numpy>=1.23.0,<2.0.0",
    "debugpy>=1.8.17",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-exporter-otlp-proto-http>=1.34.1",
    "google-genai==1.60.0"
]


[dependency-groups]
dev = [
    "babel>=2.17.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = [
    "B006", # forbid def demo(mutation = [])
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
